ingest/requirements.txt
pandas
pyarrow
google-cloud-bigquery
google-cloud-bigquery-storage
pydantic
pydantic-settings
python-dotenv
